# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "IM Chat backend API documentation.",
        "title": "IM Chat API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
        "/application/accept": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "同意好友申请",
                "parameters": [
                    {
                        "description": "同意申请请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AcceptFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "申请请求添加好友",
                "parameters": [
                    {
                        "description": "申请请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AppliSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回我收到的好友申请列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "查看全部接收到的申请",
                "responses": {
                    "200": {
                        "description": "msg=success 时 list 为申请数组",
                        "schema": {
                            "$ref": "#/definitions/models.ListAppResponse"
                        }
                    }
                }
            }
        },
        "/application/mylist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回我发出的好友申请列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "查看全部已发申请",
                "responses": {
                    "200": {
                        "description": "msg=success 时 list 为申请数组",
                        "schema": {
                            "$ref": "#/definitions/models.ListAppResponse"
                        }
                    }
                }
            }
        },
        "/application/refuse": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "拒绝好友申请",
                "parameters": [
                    {
                        "description": "拒绝申请请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RefuseFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/search/email": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "搜索邮箱查看待添加好友",
                "parameters": [
                    {
                        "description": "按邮箱搜索",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindPerson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FindEnd"
                        }
                    }
                }
            }
        },
        "/application/search/name": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "搜索名字查看待添加好友",
                "parameters": [
                    {
                        "description": "按名字模糊搜索",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindNamePerson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FindNameEnd"
                        }
                    }
                }
            }
        },
...
        "/contact/change/remark": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "修改朋友备注",
                "parameters": [
                    {
                        "description": "修改备注请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FriendRemark"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/friend/main": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "查看朋友详细信息",
                "parameters": [
                    {
                        "description": "查看请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FriendMain"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ContactMain"
                        }
                    }
                }
            }
        },
        "/contact/list": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contact"
                ],
                "summary": "查看通讯录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AcceptFriend": {
            "type": "object",
            "required": [
                "account_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                }
            }
        },
        "models.AppliSearch": {
            "type": "object",
            "required": [
                "msg",
                "send_email"
            ],
            "properties": {
                "msg": {
                    "type": "string"
                },
                "send_email": {
                    "type": "string"
                }
            }
        },
        "models.Apply": {
            "type": "object",
            "properties": {
                "from_email": {
                    "type": "string"
                },
                "from_id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "send_email": {
                    "type": "string"
                },
                "send_id": {
                    "type": "integer"
                },
                "send_name": {
                    "type": "string"
                },
                "send_pictures": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "models.Contact": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ListContact"
                    }
                }
            }
        },
        "models.ContactMain": {
            "type": "object",
            "properties": {
                "friend-picture": {
                    "type": "string"
                },
                "friend_email": {
                    "type": "string"
                },
                "friend_id": {
                    "type": "integer"
                },
                "friend_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "models.FindEnd": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "find": {
                    "$ref": "#/definitions/models.FindMiddle"
                }
            }
        },
        "models.FindMiddle": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pictures": {
                    "type": "string"
                }
            }
        },
        "models.FindNameEnd": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "find": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FindNameMiddle"
                    }
                }
            }
        },
        "models.FindNameMiddle": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pictures": {
                    "type": "string"
                }
            }
        },
        "models.FindNamePerson": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.FindPerson": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.FriendMain": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                }
            }
        },
        "models.FriendRemark": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "models.ListAppResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Apply"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.ListContact": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "friend_name": {
                    "type": "string"
                },
                "friend_picture": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.RefuseFriend": {
            "type": "object",
            "required": [
                "account_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}